+++ src/main/java/twilightforest/entity/EntityTFGoblinKnightUpper.java
@@ -90,7 +90,7 @@ public class EntityTFGoblinKnightUpper extends EntityMob {
     protected void entityInit()
     {
         super.entityInit();
-        dataWatcher.addObject(DATA_EQUIP, Byte.valueOf((byte)0));
+        dataWatcher.addObject(DATA_EQUIP, (byte)0);
     }
 	
     public boolean hasArmor()
@@ -105,11 +105,11 @@ public class EntityTFGoblinKnightUpper extends EntityMob {
     	
         if (flag)
         {
-            dataWatcher.updateObject(DATA_EQUIP, Byte.valueOf((byte) (otherFlags | 1)));
+            dataWatcher.updateObject(DATA_EQUIP, (byte) (otherFlags | 1));
         }
         else
         {
-            dataWatcher.updateObject(DATA_EQUIP, Byte.valueOf((byte)otherFlags));
+            dataWatcher.updateObject(DATA_EQUIP, (byte)otherFlags);
         }
     }
     
@@ -125,11 +125,11 @@ public class EntityTFGoblinKnightUpper extends EntityMob {
     	
         if (flag)
         {
-            dataWatcher.updateObject(DATA_EQUIP, Byte.valueOf((byte) (otherFlags | 2)));
+            dataWatcher.updateObject(DATA_EQUIP, (byte) (otherFlags | 2));
         }
         else
         {
-            dataWatcher.updateObject(DATA_EQUIP, Byte.valueOf((byte)otherFlags));
+            dataWatcher.updateObject(DATA_EQUIP, (byte)otherFlags);
         }
     }
